# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DevDex API - Stack Overflow Style Q&A Platform
  description: API documentation for DevDex - A developer Q&A platform with authentication, questions, answers, voting, tags, and payment integration
  version: 2.0.0
  contact:
    name: DevDex Support
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only cookie
  schemas:
...
    Question:
      type: object
      required:
        - title
        - desc
        - username
        - userId
      properties:
        _id:
          type: string
          readOnly: true
        title:
          type: string
          description: Question title
        desc:
          type: string
          description: Question body/description
        photo:
          type: string
          description: Optional image URL
        username:
          type: string
        userId:
          type: string
        categories:
          type: array
          items:
            type: string
          description: Tags associated with the question
        votes:
          type: integer
          default: 0
        upvoters:
          type: array
          items:
            type: string
        downvoters:
          type: array
          items:
            type: string
        acceptedAnswerId:
          type: string
          nullable: true
        views:
          type: integer
          default: 0
        answerCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Tag:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        name:
          type: string
          description: Unique tag name (lowercase)
        description:
          type: string
        questionCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
    Comment:
      type: object
      required:
        - comment
        - author
        - postId
        - userId
      properties:
        _id:
          type: string
          readOnly: true
        comment:
          type: string
        author:
          type: string
        postId:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
    VoteRequest:
      type: object
      required:
        - voteType
        - userId
      properties:
        voteType:
          type: integer
          enum: [1, -1]
          description: 1 for upvote, -1 for downvote
        userId:
          type: string
...
paths:
...
  /api/posts:
    get:
      tags: [Questions]
      summary: Get all questions with filtering and sorting
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search in title and description
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, active, unanswered, votes]
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag name
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
  /api/posts/create:
    post:
      tags: [Questions]
      summary: Create a new question
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, desc, username, userId]
              properties:
                title:
                  type: string
                desc:
                  type: string
                photo:
                  type: string
                username:
                  type: string
                userId:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /api/posts/{id}:
    get:
      tags: [Questions]
      summary: Get question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'

    put:
      tags: [Questions]
      summary: Update question
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                desc:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Question updated

    delete:
      tags: [Questions]
      summary: Delete question and all associated answers
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question deleted
  /api/posts/{id}/vote:
    put:
      tags: [Questions]
      summary: Vote on a question
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '200':
          description: Vote registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /api/posts/{id}/view:
    put:
      tags: [Questions]
      summary: Increment view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: View count incremented
  /api/posts/{id}/accept/{answerId}:
    put:
      tags: [Questions]
      summary: Accept an answer
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: answerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Answer accepted
  /api/posts/user/{userId}:
    get:
      tags: [Questions]
      summary: Get questions by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User's questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
  /api/posts/tag/{tagName}:
    get:
      tags: [Questions]
      summary: Get questions by tag
      parameters:
        - name: tagName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Questions with tag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'

  # Answer Endpoints
...
  /api/tags:
    get:
      tags: [Tags]
      summary: Get all tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      tags: [Tags]
      summary: Create or update tag
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Tag created/updated
...
  /api/tags/search:
    get:
      tags: [Tags]
      summary: Search tags
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Matching tags
  /api/tags/{name}:
    get:
      tags: [Tags]
      summary: Get tag by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'

  # Comment Endpoints
...
  /api/comments/post/{postId}:
    get:
      tags: [Comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
...